refactor(library/utilitiesByType): moves "/string.ts" to "/string/concatenated.ts"
