+++ src/library/StringSubset.ts
@@ -5,7 +5,7 @@ import type {
 import {
   type StringLike,
   concatenated,
-} from '@/library/utilitiesByType/string';
+} from '@/library/utilitiesByType/string/concatenated';
 
 /**
  * When an open-ended `string` is too permissive, extend this class and provide a means to instantiate some subset
