+++ src/library/utilitiesByType/string/concatenated.ts
